Create AM_Mod.m
Simulate an amplitude modulation (AM) system with all input,  carrier   and output signals.  Say the input signal is a cosine wave  with amplitude  as 2V and frequency as 1000Hz. The carrier signal is also a cosine wave with amplitude as 5V and frequency as 10KHz.  The modulation degree   is 0.5, and the initial phases of all cosine wave are 0. (Recall Nyquist sampling theorem to avoid distortion i.e. under sampling